requirements.txt
speechbrain>=0.5.14
streamlit>=1.25.0
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
plotly>=5.0.0
python-json-logger>=2.0.0
datasets>=2.14.0
pyannote.audio>=3.1.0
python-dotenv>=0.19.0